test: di module에 InjectorTest 구현
