+++ di/build.gradle.kts
@@ -7,5 +7,14 @@ java {
     targetCompatibility = JavaVersion.VERSION_11
 }
 dependencies {
+    implementation("javax.inject", "javax.inject", "1")
     implementation(kotlin("reflect"))
+    testImplementation("org.junit.jupiter", "junit-jupiter", "5.8.2")
+    testImplementation("org.assertj", "assertj-core", "3.22.0")
+}
+
+tasks {
+    test {
+        useJUnitPlatform()
+    }
 }
